# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.AppointmentListVO:
    properties:
      appointment_date:
        type: string
      appointment_no:
        type: string
      appointment_time:
        type: string
      can_cancel:
        type: boolean
      can_checkin:
        type: boolean
      department_name:
        type: string
      doctor_avatar:
        type: string
      doctor_name:
        type: string
      id:
        type: integer
      patient_name:
        type: string
      period_name:
        type: string
      status:
        type: string
      status_name:
        type: string
    type: object
  model.AppointmentVO:
    properties:
      appointment_date:
        type: string
      appointment_no:
        type: string
      appointment_time:
        type: string
      can_cancel:
        description: 是否可取消
        type: boolean
      can_checkin:
        description: 是否可签到
        type: boolean
      cancel_reason:
        type: string
      cancelled_at:
        type: string
      checked_in_at:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      department_id:
        type: integer
      department_name:
        type: string
      doctor_avatar:
        type: string
      doctor_id:
        type: integer
      doctor_name:
        type: string
      doctor_title:
        type: string
      id:
        type: integer
      patient_id:
        type: integer
      patient_name:
        type: string
      period:
        type: string
      period_name:
        type: string
      slot_number:
        type: integer
      status:
        type: string
      status_name:
        type: string
      symptom:
        type: string
      user_id:
        type: integer
    type: object
...
  model.LoginLogVO:
    properties:
      created_at:
        type: string
      device:
        type: string
      id:
        type: integer
      ip:
        type: string
      location:
        type: string
      login_type:
        type: string
      message:
        type: string
      status:
        type: integer
      user_id:
        type: integer
      user_type:
        type: string
      username:
        type: string
    type: object
...
  model.OperationLogVO:
    properties:
      action:
        type: string
      admin_id:
        type: integer
      admin_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      ip:
        type: string
      latency:
        type: integer
      method:
        type: string
      module:
        type: string
      path:
        type: string
      status:
        type: integer
    type: object
...
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.CancelAppointmentRequest:
    properties:
      reason:
        maxLength: 256
        minLength: 2
        type: string
    required:
    - reason
    type: object
...
  service.CreateAppointmentRequest:
    properties:
      idempotent_token:
        type: string
      patient_id:
        minimum: 1
        type: integer
      schedule_id:
        minimum: 1
        type: integer
      symptom:
        maxLength: 512
        type: string
    required:
    - idempotent_token
    - patient_id
    - schedule_id
    type: object
...
  service.UpdateAppointmentStatusRequest:
    properties:
      remark:
        maxLength: 256
        type: string
      status:
        enum:
        - pending
        - checked_in
        - completed
        - cancelled
        - missed
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@huaan-medical.com
    name: API Support
  description: 华安医疗门诊预约系统后端API接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 华安医疗预约系统 API
  version: "1.0"
paths:
...
  /api/admin/appointments:
    get:
      consumes:
      - application/json
      description: 分页查询所有预约，支持日期范围、状态、关键词筛选
      parameters:
      - description: 页码
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        required: true
        type: integer
      - description: 开始日期（YYYY-MM-DD）
        in: query
        name: start_date
        type: string
      - description: 结束日期（YYYY-MM-DD）
        in: query
        name: end_date
        type: string
      - description: 预约状态
        in: query
        name: status
        type: string
      - description: 搜索关键词（预约编号、患者姓名、医生姓名）
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/model.AppointmentVO'
                        type: array
                    type: object
              type: object
      security:
      - BearerAdmin: []
      summary: 查询预约列表（管理后台）
      tags:
      - 预约管理
  /api/admin/appointments/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定预约的详细信息
      parameters:
      - description: 预约ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AppointmentVO'
              type: object
      security:
      - BearerAdmin: []
      summary: 获取预约详情（管理后台）
      tags:
      - 预约管理
    put:
      consumes:
      - application/json
      description: 管理员更新预约状态
      parameters:
      - description: 预约ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateAppointmentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAdmin: []
      summary: 更新预约状态（管理后台）
      tags:
      - 预约管理
  /api/admin/appointments/export:
    get:
      consumes:
      - application/json
      description: 导出预约数据为CSV文件
      parameters:
      - description: 开始日期（YYYY-MM-DD）
        in: query
        name: start_date
        type: string
      - description: 结束日期（YYYY-MM-DD）
        in: query
        name: end_date
        type: string
      - description: 预约状态
        in: query
        name: status
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV文件
          schema:
            type: file
      security:
      - BearerAdmin: []
      summary: 导出预约数据（管理后台）
      tags:
      - 预约管理
...
  /api/admin/logs/login:
    get:
      consumes:
      - application/json
      description: 分页查询登录日志，支持关键词/状态/日期筛选
      parameters:
      - description: 页码
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        required: true
        type: integer
      - description: 用户类型 user/admin
        in: query
        name: user_type
        type: string
      - description: 状态 0失败 1成功
        in: query
        name: status
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 开始日期（YYYY-MM-DD）
        in: query
        name: start_date
        type: string
      - description: 结束日期（YYYY-MM-DD）
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/model.LoginLogVO'
                        type: array
                    type: object
              type: object
      security:
      - BearerAdmin: []
      summary: 查询登录日志（管理后台）
      tags:
      - 登录日志
  /api/admin/logs/operation:
    get:
      consumes:
      - application/json
      description: 分页查询操作日志，支持关键词/模块/日期筛选
      parameters:
      - description: 页码
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        required: true
        type: integer
      - description: 管理员ID
        in: query
        name: admin_id
        type: integer
      - description: 模块
        in: query
        name: module
        type: string
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 开始日期（YYYY-MM-DD）
        in: query
        name: start_date
        type: string
      - description: 结束日期（YYYY-MM-DD）
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/model.OperationLogVO'
                        type: array
                    type: object
              type: object
      security:
      - BearerAdmin: []
      summary: 查询操作日志（管理后台）
      tags:
      - 操作日志
...
  /api/appointments:
    get:
      consumes:
      - application/json
      description: 查询当前用户的所有预约
      parameters:
      - description: 预约状态（pending/checked_in/completed/cancelled/missed）
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.AppointmentListVO'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: 查询用户预约列表
      tags:
      - 预约
    post:
      consumes:
      - application/json
      description: 用户创建新预约
      parameters:
      - description: 预约信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AppointmentVO'
              type: object
      security:
      - Bearer: []
      summary: 创建预约
      tags:
      - 预约
  /api/appointments/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定预约的详细信息
      parameters:
      - description: 预约ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AppointmentVO'
              type: object
      security:
      - Bearer: []
      summary: 获取预约详情
      tags:
      - 预约
  /api/appointments/{id}/cancel:
    put:
      consumes:
      - application/json
      description: 用户取消预约（就诊当天不可取消）
      parameters:
      - description: 预约ID
        in: path
        name: id
        required: true
        type: integer
      - description: 取消原因
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CancelAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: 取消预约
      tags:
      - 预约
  /api/appointments/{id}/checkin:
    post:
      consumes:
      - application/json
      description: 用户在预约时间范围内签到
      parameters:
      - description: 预约ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: 预约签到
      tags:
      - 预约
...
securityDefinitions:
  Bearer:
    description: 请在Token前加上"Bearer "前缀
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"